fix: address CodeRabbit third-pass findings on WS-03 PR
- Mirror the friend_requests.message length CHECK in the Drizzle schema
  (sourced from the same FRIEND_REQUEST_MESSAGE_MAX_LENGTH constant so
  a future bump lands at HTTP + DB + docs together).
- Name every friend_requests / friendships / user_blocks CHECK
  constraint in the migration so the drizzle-kit snapshot's inline
  CHECKs can reference the same identifiers and the DB error messages
  stay meaningful.
- Factor isUniqueViolation / extractPgConstraint into
  apps/api/src/shared/pg-errors.ts; rooms and friends repositories
  re-export from there so the translation of 23505 into domain codes
  stays a single point of truth.
- Normalize blank/whitespace-only friend-request messages to NULL before
  insert so the DB keeps a single canonical "no message" representation.
- Use UsernameSchema (instead of a plain Type.String()) on the response
  side of FriendRequestPublicSchema so request + response validation
  stay aligned if username rules evolve.
